backend/requirements.txt
flask
flask-cors
flask-limiter
openai
faiss-cpu
httpx
python-dotenv